PP-6000 Move newSecondFactorPasscode from UserServices to new class
Move the newSecondFactorPasscode(…) method from UserServices to
the new ExistingUserOtpDispatcher class. No functionality changes.
